+++ utils/test/feedback/family_tree_test.exs
@@ -1,0 +1,5 @@
+defmodule Utils.Feedback.FamilyTreeTest do
+  use Utils.FeedbackCase,
+    module: Utils.Feedback.FamilyTree,
+    file: "../exercises/family_tree.livemd"
+end
